the DF is ready for handling histos
