+++ python/histo_classes.py
@@ -1,0 +1,37 @@
+import ROOT
+from   ROOT import TFile
+from   ROOT import gDirectory
+import sys
+
+
+def GetKeyNames( self, dir = "" ):
+    '''some magic to get all directories within dir'''
+    self.cd(dir)
+    return [key.GetName() for key in gDirectory.GetListOfKeys()]
+TFile.GetKeyNames = GetKeyNames
+
+
+class HistogramFile(object):
+    '''
+    basic interface to a root file
+    '''
+    def __init__(self, filename):
+        self.filename = filename
+
+    def __enter__(self):
+        '''http://effbot.org/zone/python-with-statement.htm'''
+        self.file = ROOT.TFile.Open(self.filename, 'read')
+        return self
+
+    def __exit__(self, exception_type, exception_value, traceback):
+        self.file.Close()
+
+    def get_histogram(self, name):
+        """Return the histogram identified by name from the file.
+        """
+        # The TFile::Get() method returns a pointer to an object stored in a ROOT file.
+        hist = self.file.Get(name)
+        if hist:
+            return hist
+        else:
+            raise RuntimeError('Unable to retrieve histogram named {0} from {1}'.format(name, self.filename))
